added support for IP agnostic dogstatsd
